need to cast x and y resize values for IRIX
git-svn-id: file:///home/users/tmap/svn/repos/ferret/trunk@8011 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ fer/ccr/resize_animation_window.c
@@ -22,7 +22,8 @@ float *y;
 
   ws  = OPEN_WSID (*ws_id);
 
-  XResizeWindow (ws->dpy,ws->win,x,y);
+  /* 3/02 *kob* need to cast x and y for IRIX */
+  XResizeWindow (ws->dpy,ws->win,(unsigned int)x,(unsigned int)y);
 
  
 }
